requirements.txt
asgiref==3.8.1
autopep8==2.3.2
Django==5.2.1
django-dirtyfields==1.9.7
django-silk==5.3.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
gprof2dot==2025.4.14
inflection==0.5.1
packaging==25.0
pillow==11.2.1
pycodestyle==2.13.0
PyJWT==2.9.0
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.13.2
uritemplate==4.1.1
